Added robustness to requesting data.
The device always sends temp then hum, temp then hum, etc.  It's possible to
get out of sync with the device and expect temp but get hum.  This patch checks
for temperature first, which is three bytes and if it doesn't get three bytes,
it throws out the result and retries the request in hopes to get back in sync.

This situation should only occur on startup, not arbitrarily during a run.