Added check for valid data before returning.
Make sure the temperature are within spec; +/- 200 for temperature and 0..100
for humidity.  If the data is not correct, retry twice.

It appears that there is some initial data in the buffer that is not temp or
humidity when a computer is first booted up.  This retry mechanism clears out
the buffer until there is reasonable data in the pipe.